backend/pyproject.toml
[project]
name = "polygents"
version = "0.1.0"
description = "Multi-agent collaboration framework"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pyyaml>=6.0",
    "watchfiles>=0.20.0",
    "claude-agent-sdk>=0.1.0",
    "anyio>=4.0.0",
    "pydantic>=2.0.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
]
